// schemas/web-checkbox.ts
export interface CheckboxProps {
  className: string;
  checked: string;
  disabled: boolean;
  indeterminate: string;
  ariaChecked: string;
  ariaLabel: string;
}

export interface CheckboxFieldProps {
  wrapperClassName: string;
  className: string;
  checked: string;
  disabled: boolean;
  text: string;
}